# Conflicts:
#	deepchecks/tabular/checks/distribution/train_test_feature_drift.py
#	deepchecks/tabular/checks/distribution/train_test_label_drift.py
#	deepchecks/tabular/checks/distribution/train_test_prediction_drift.py
#	deepchecks/utils/distribution/drift.py
#	deepchecks/vision/checks/distribution/image_property_drift.py
#	deepchecks/vision/checks/distribution/train_test_label_drift.py
#	deepchecks/vision/checks/distribution/train_test_prediction_drift.py
#	spelling-allowlist.txt

+++ deepchecks/utils/single_sample_metrics.py
@@ -62,7 +62,7 @@ def per_sample_cross_entropy(y_true: np.array, y_pred: np.array, eps=1e-15):
         )
 
     # clip and renormalization y_pred
-    y_pred = y_pred.clip(eps, 1 - eps)
+    y_pred = y_pred.astype('float').clip(eps, 1 - eps)
     y_pred /= y_pred.sum(axis=1)[:, np.newaxis]
 
     return -(transformed_labels * np.log(y_pred)).sum(axis=1)
